Brudite-SES/requirements.txt
asgiref==3.7.2
Django==4.2.5
django-restframework==0.0.1
django-storages==1.14
djangorestframework==3.14.0
et-xmlfile==1.1.0
numpy==1.26.0
openpyxl==3.1.2
pandas==2.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
